Extract scoring module (CAS engine) from the monolith (split PR 3/N)
The largest single-PR extraction so far: moves the Combined Affinity
Score engine — find_clusters plus its 10 per-component weight helpers
and calcCombinedAffinityPvalue — into tfbs_footprinter3/scoring.py.

Extracted:
- calcCombinedAffinityPvalue (pinned by tests/test_cas_pvalue.py)
- find_clusters (the CAS orchestrator)
- eqtl_overlap_likelihood
- eqtls_weights_summing
- cage_correlations_summing_preparation
- cage_correlations_summing
- cage_weights_summing
- atac_weights_summing
- metacluster_weights_summing
- gerp_weights_summing
- cpg_weights_summing

The new module's only internal dependency is
tfbs_footprinter3.io_utils (overlap_range, distance_solve), which is
clean — no circular import with the monolith.

Also drops the now-obsolete ##-prefixed earlier versions of each
function (atac/metacluster/gerp/cpg weights_summing all had older
commented-out implementations). Those removals are behavior-neutral
dead-code cleanup; the active function bodies were copied verbatim.

Monolith: 3024 -> 2551 lines (-473). All 21 unit tests green, ruff
clean, smoke verifies every re-exported name still resolves.
